% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{add_to_corpus}
\alias{add_to_corpus}
\title{Add new files to a corpus
Add new files to a corpus without reaggregating everything}
\usage{
add_to_corpus(old_corpus_path, files_updated_path, ENGINE)
}
\arguments{
\item{old_corpus_path}{path to old corpus. List}

\item{files_updated_path}{directory with some new pdfs.}

\item{ENGINE}{similar to engine argument in readPDF from 'tm' package.
Function to read pdf into environment, either pdf_text or pdf_ocr_text, depending on whether image or not.}
}
\description{
Add new files to a corpus
Add new files to a corpus without reaggregating everything
}
\author{
Manuel Betin, Umberto Collodel
}
