% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{aggregate_corpus}
\alias{aggregate_corpus}
\title{Aggregate pdf files into list of characters}
\usage{
aggregate_corpus(path_files, ENGINE = "pdf_text", only_files = F)
}
\arguments{
\item{path_files}{the path of the directory with the files}

\item{ENGINE}{similar to engine argument in readPDF from 'tm' package. 
Function to use to read pdf into environment, either pdf_text or pdf_ocr_text.}

\item{only_files}{T/F whether to include in the list only the content or also
the metadata of the pdf file}
}
\value{
A list containing the content of each document
}
\description{
function that takes the path of the directory and load all
the pdfs of the directory into a list in order to further perform the text
mining
}
\author{
Manuel Betin
}
