\name{assoc_words}
\alias{assoc_words}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
from the table of ngrams generated by ngram_clean() find the most common associated words by typology
}
\description{
from the table of ngrams generated by ngram_clean() find the most common associated words by typology
}
\usage{
assoc_words(clean_ngram, type_word, min_occurence = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{clean_ngram}{
output from ngram_clean()
}
  \item{type_word}{
the name of the category to analyse (check outptu of ngram_clean() to see the available list)
}
  \item{min_occurence}{
only display the words with at least this specified number of occurence
}
}

\value{
a tibble with ngrams in rows, number of occurence and dummies for the typologies 
}
\author{
Manuel Betin
}