\name{binary_freq_trans}
\alias{binary_freq_trans}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 transform table from Number of occurence to binary variables
}
\description{
 transform table from Number of occurence to binary variables
}
\usage{
binary_freq_trans(table_N_occurence)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{table_N_occurence}{
 output of tf_vector()
}
}
\value{
the function returns the binary frequency transformation of the dataframe.

}
\author{
Manuel Betin
}
