\name{CCDr-package}
\alias{CCDr-package}
\alias{CCDr}
\docType{package}
\title{
\packageTitle{CCDr}
}
\description{
\packageDescription{CCDr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CCDr}
\packageIndices{CCDr}
 
}
\author{
\packageAuthor{CCDr}

Maintainer: \packageMaintainer{CCDr}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{ Text Mining, Economic crisis, Natural lenguage processing}