% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.corpus.countryupdate}
\alias{ccdr.corpus.countryupdate}
\title{update corpus for a specific country}
\usage{
ccdr.corpus.countryupdate(iso3, historical_vintage, new_vintage)
}
\arguments{
\item{iso3}{the iso3 code of the country of interest}

\item{historical_vintage}{object of class corpusTM to update}

\item{new_vintage}{object of class corpusTM with the corpus of the new update}
}
\description{
update corpus for a specific country
}
\author{
Manuel
}
