% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.r
\name{ccdr.fasttext.predict}
\alias{ccdr.fasttext.predict}
\title{predict the label foreach sentences in the corpus}
\usage{
ccdr.fasttext.predict(
  mycorpus,
  path_rawtext = "../../inst/rawdata/rawtext",
  path_output = "../../inst/results",
  min_sentence_lengh = 20,
  rm_stopwords = T,
  model,
  k,
  th
)
}
\arguments{
\item{mycorpus}{A list containg the text to be analyzed (output of ccdr.corpus() 
from the CCRr package)}

\item{path_rawtext}{Path to the folder that contains the text tokenized in sentences}

\item{path_output}{Path where to store the predictions}

\item{rm_stopwords}{T/F specified whether you want to remove digits and stopwords in
the text for prediction.}

\item{min_sentence_length}{The threshold for the minimum number of words for a sentence
to be considered}

\item{...}{Arguments to the fasttext_interface (model, k, th) from the package fastText}
}
\description{
Predict the label of each sentences from a corpus of $
documents provided
}
\author{
Manuel Betin
}
