% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.r
\name{ccdr.fasttext.prediction.spline}
\alias{ccdr.fasttext.prediction.spline}
\title{correct the missing years in the sample}
\usage{
ccdr.fasttext.prediction.spline(mypredictions)
}
\description{
interpolate missing years with linear 
approximation
}
