% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.r
\name{ccdr.lexicon.exportcsv}
\alias{ccdr.lexicon.exportcsv}
\title{export dictionary of words in csv}
\usage{
ccdr.lexicon.exportcsv(myvars = NULL, path)
}
\arguments{
\item{myvars}{a vector with the names of the categories to export}

\item{path}{the path to the folder where the dicitonary is exported}
}
\description{
export the dictionary
}
\author{
Manuel Betin
}
