% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.r
\name{ccdr.lexicon}
\alias{ccdr.lexicon}
\title{Lexicon of economic crisis}
\usage{
ccdr.lexicon()
}
\value{
a list of categories with the corresponding words
 associated
}
\description{
Provide the lexicon of economic crisis
}
