% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.sentences}
\alias{ccdr.sentences}
\title{From a corpus (collection documents), returns the sentences where keyword detected
and respective keyword}
\usage{
ccdr.sentences(corpus, keyword_list)
}
\arguments{
\item{corpus}{list with collection documents}

\item{keyword_list}{character vector with categories from key_words_crisis function}
}
\value{
Tibble with two columns: sentences containing keyword and respective keyword.
}
\description{
Function that simplifies checking validity indexes and potential problems in keywords
}
\author{
Manuel Betin, Umberto Collodel
}
