% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tf.fig.plotIndex}
\alias{ccdr.tf.fig.plotIndex}
\title{produce the time serie for the selected crisis}
\usage{
ccdr.tf.fig.plotIndex(
  CCDB_y,
  myvar,
  type = "mean",
  year_window = c(1950, 2022)
)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB database}

\item{myvar}{the name of the crisis of interest}

\item{type}{the type of statistics, N=Number of countries with a positive term frequency
, mean=average term frequency, perc=percentage of countries with a positive term frequency}

\item{years_window}{a vector with first and last date to plot.}
}
\description{
figure that plot the time serie of the crisis of interest
}
\author{
Manuel Betin
}
