% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.ctrycoverage}
\alias{ccdr.tfs.fig.ctrycoverage}
\title{plot the number of countries in the sample by year}
\usage{
ccdr.tfs.fig.ctrycoverage(
  CCDB_y,
  type_classif = "income_group",
  year_window = c(1950, 2022),
  const_sample_since = 2022
)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dataset}

\item{type_classif}{the type of country classification for the grouping variable}

\item{year_window}{a vector with the first and last year of interest}

\item{const_sample_since}{the date before which only members countries are available in 
the sample}
}
\description{
time series with the number of country in the sample by income_group or region
according to the world bank classification
}
\author{
Manuel Betin
}
