% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.eventstudy}
\alias{ccdr.tfs.fig.eventstudy}
\title{plot the crisis index for the selected country}
\usage{
ccdr.tfs.fig.eventstudy(CCDB_y, iso3c, myvar)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dabaset}

\item{iso3c}{the iso3 code of the country of interest}

\item{myvar}{the name of the indicator to plot}
}
\description{
plot the selected indicators for the country of interest
}
\author{
Manuel Betin
}
