% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.eventstudyYES}
\alias{ccdr.tfs.fig.eventstudyYES}
\title{plot the number of in crisis by typology}
\usage{
ccdr.tfs.fig.eventstudyYES(CCDB_y, iso3c)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dataset}

\item{iso3c}{the iso3 code of the country of interest}
}
\description{
plot the time series of SGS,FGS and risk free crisis for the 
selected country
 SGS crisis corresponds to sovereign crisis with expectation regime
shift, FGS to sovereign crisis without expectation regime shifts and risk free 
recessions crisis without sovereign debt concerns.
}
\author{
Manuel Betin
}
