% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.network}
\alias{ccdr.tfs.fig.network}
\title{plot the network of crisis}
\usage{
ccdr.tfs.fig.network(CCDB_y, min_year = 2018, max_year = 2022, min_corr = 0.2)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dataset}

\item{min_year}{the first year of the window}

\item{max_year}{the last year of the window}

\item{min_corr}{the minimum correlation to draw the edges in the network}
}
\description{
plot the network of the correlations matrix of crisis for the 
specific window of periods.
}
\author{
Manuel Betin
}
