% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.radiography}
\alias{ccdr.tfs.fig.radiography}
\title{plor the variations of indices around crisis by income group}
\usage{
ccdr.tfs.fig.radiography(CCDB_y, classif = "all", vars, years_crisis, window)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dabaset}

\item{classif}{all/income_group/region}

\item{vars}{a vector with the names of the categories of interest}

\item{years_crisis}{a vector of dates for the crisis of interest}

\item{window}{a vector with the range of periods around the crisis event}
}
\description{
plot the selected indicators in a window around crisis period for the 
selected time span and the income group of interest
}
\author{
Manuel Betin
}
