% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.tsfreq}
\alias{ccdr.tfs.fig.tsfreq}
\title{plot the time serie of the index for the selected statistic}
\usage{
ccdr.tfs.fig.tsfreq(CCDB_y, myvar, type = "perc", year_window = c(1950, 2022))
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB dabaset}

\item{type}{N/mean/perc}

\item{year_window}{a vector with the range of periods around the crisis event}

\item{myvars}{the names of the category of interest}
}
\description{
plot the total number of countries (N), the percentage (perc),
or the average term frequency (mean) for the select indicators by income 
group.
}
\author{
Manuel Betin
}
