% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.normalize}
\alias{ccdr.tfs.normalize}
\title{nomarlize term frequencies}
\usage{
ccdr.tfs.normalize(mydata, var2normalize)
}
\arguments{
\item{mydata}{a vintage of the CCDB dataset}

\item{var2normalize}{a vector of column names for which the normalisation is
performed.}
}
\description{
normalize the variable of interest by removing the
country specific means and dividing by the standard deviations
when all serie is zero then the value is set to zero.
}
\author{
Manuel Betin
}
