% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.typologycrisis}
\alias{ccdr.tfs.typologycrisis}
\title{compute typology of crisis}
\usage{
ccdr.tfs.typologycrisis(mydata)
}
\arguments{
\item{mydata}{a vintage of the CCDB database}
}
\description{
compute the typology of crisis in the spirit of Betin 2022
that discriminate FGS and SGS crisis
}
\author{
Manuel Betin
}
