% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.updateyears}
\alias{ccdr.tfs.updateyears}
\title{append old vintage and new vintage updated for new data}
\usage{
ccdr.tfs.updateyears(old_vintage, updated_years)
}
\arguments{
\item{old_vintage}{data.frame of the old vintage of the database for which new years are available}

\item{updated_years}{data.frame of the new variables updated.}
}
\description{
update the old vintage of the database with the the new dataset
for the updated years
}
\author{
Manuel Betin
}
