% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.collect}
\alias{ccdr.transcripts.collect}
\title{collect transcripts identified as belonging to a specific category}
\usage{
ccdr.transcripts.collect(mycorpus, keyword_list)
}
\arguments{
\item{mycorpus}{a list of documents containing the reports of interest}

\item{keyword_list}{a vector with the names of the categories to search for
see ccdr.lexicon() %>% names() for details about available categories}
}
\value{
tibble() with the transcripts, the keyword detected, the category of
crisis and the identifier of the document.
}
\description{
the function extract the paragraphs around the keywords that
correspond to the categories of interest selected in key_word_list
}
\author{
Manuel Betin
}
