% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.r
\name{ccdr.transcripts.compile}
\alias{ccdr.transcripts.compile}
\title{Get transcripts for a specific variable}
\usage{
ccdr.transcripts.compile(myvar, ctries, path_corpus, path_export)
}
\arguments{
\item{myvar}{the name of the category in the dictionary that will be extracted}

\item{ctries}{a vector of countries of which extract the transcripts}

\item{path_corpus}{the path of the directory where the corpus for each country
are located}

\item{path_export}{the path of the directory where to save the specific transcripts}
}
\description{
use the CCDB dictionary to extract and label each sentence of 
the country reports.
}
\author{
Manuel Betin
}
