% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.r
\name{ccdr.transcripts.fasttextsample}
\alias{ccdr.transcripts.fasttextsample}
\title{aggregate the transcript into training and test sample}
\usage{
ccdr.transcripts.fasttextsample(
  path_transcripts,
  training_size = 0.7,
  min_char = 30,
  min_word = 15,
  sentiment_threshold = 5,
  neutral_sentiment = c(-0.25, 0.25),
  multi_label = T,
  rm_stop_words = T,
  sentiment_corpus = NULL
)
}
\arguments{
\item{path_transcripts}{a vector containing the path to the ccdb_transcripts
(output of the function ccdr.transcripts.collect())}

\item{training_size}{proportion of the sample allocated to training sample}

\item{min_char}{Threshold for minimum number of character for sentences to be valid}

\item{min_word}{threshold for minimum number of words for sentences to be valid}

\item{sentiment_threshold}{only keep sentences with negative sentiment below the threshold.}

\item{neutral_sentiment}{vector with the bounds for the definition of neutral sentiment}

\item{multi_label}{allow several labels per sentences.}

\item{rm_stop_words}{T/F remove stopwords}

\item{sentiment_corpus}{a dataset with the classification of negative and positive words, the dataset should 
have at least two colonnes word and sentiment.}
}
\description{
create a training and test sample of text compatible with the 
fasttext algorithm, for the supervized  text recognition model
}
\author{
Manuel Betin
}
