% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.get_nearby_words}
\alias{ccdr.transcripts.get_nearby_words}
\title{check the frequency of nearby words}
\usage{
ccdr.transcripts.get_nearby_words(word_dt, myword, n_min = 10)
}
\description{
provide a table with the frequencies of nearby words
}
