% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.plot.KeywordsFreq}
\alias{ccdr.transcripts.plot.KeywordsFreq}
\title{plot the keyword frequency}
\usage{
ccdr.transcripts.plot.KeywordsFreq(mytopkeywords, mywords)
}
\description{
plot the keyword frequency
}
\author{
Manuel Betin
}
