% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.preprocess.FinCrisisWordsfreq}
\alias{ccdr.transcripts.preprocess.FinCrisisWordsfreq}
\title{preprocess the category financial crisis}
\usage{
ccdr.transcripts.preprocess.FinCrisisWordsfreq(mytokens)
}
\description{
group similar words to make sure that the frequency of words
take into account the different expressions
}
\author{
Manuel Betin
}
