% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.preprocess.Keywordsfreq}
\alias{ccdr.transcripts.preprocess.Keywordsfreq}
\title{tokenize and analyze transcripts with tidyr}
\usage{
ccdr.transcripts.preprocess.Keywordsfreq(mytokens, myvar)
}
\description{
tokenize the transcripts and remove the stop words
}
\author{
Manuel Betin
}
