% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{ccdr.transcripts.tokenize}
\alias{ccdr.transcripts.tokenize}
\title{tokenize and analyze transcripts with tidyr}
\usage{
ccdr.transcripts.tokenize(mytranscriptsdt)
}
\description{
tokenize the transcripts and remove the stop words
}
\author{
Manuel Betin
}
