% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{check_diff_pdfs_urls}
\alias{check_diff_pdfs_urls}
\title{Check difference between pdfs downloaded and original urls
Check difference between pdfs downloaded and original urls}
\usage{
check_diff_pdfs_urls(path, urls_data)
}
\arguments{
\item{path}{path to all individual units subdirectories.}

\item{urls_data}{dataframe. original urls, must have at least a column named "name_file".}
}
\description{
Check difference between pdfs downloaded and original urls
Check difference between pdfs downloaded and original urls
}
\author{
Manuel Betin, Umberto Collodel
}
