% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{check_gaps_pdfs}
\alias{check_gaps_pdfs}
\title{check if there are gaps greater than one year in the documents}
\usage{
check_gaps_pdfs(general_path = "/Volumes/Elements/IMF documents/",
  only_large = TRUE)
}
\arguments{
\item{general_path}{path with one subdirectory per individual and for each of them a files subdirectory with pdfs}

\item{only_large}{Logical. Consider only gaps greater than two years. Default is TRUE.}
}
\value{
a nested list: for every individual the number of gaps and a dataframe with the details.
}
\description{
check if there are gaps greater than one year in the documents
}
\author{
Manuel Betin, Umberto Collodel
}
