\name{clean_text}
\alias{clean_text}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
clean a string of character by removing special characters, tabulations, very short paragraphs...
}
\description{
clean a string of character by removing special characters, tabulations, very short paragraphs...

}
\usage{
clean_text(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
a string of characters to clean
}
}
\value{
a string of characters cleanned
}

\author{
Manuel Betin
}