% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimilarity.R
\name{cosim}
\alias{cosim}
\title{Compute cosinus similarity of two vectors
Measure the distance between two vectors using cosinus similarity
formula}
\usage{
cosim(tf_data, vec1, vec2)
}
\arguments{
\item{tf_data}{a tibble with documents in rows and number 
of occurence of each type of crisi in columns}

\item{vec1}{the name of the column 1 to analyse}

\item{vec2}{the name of the column 2 to analyse}
}
\value{
A number providing the similarity between the two vectors
between the two vectors
}
\description{
Compute cosinus similarity of two vectors
Measure the distance between two vectors using cosinus similarity
formula
}
\author{
Manuel Betin
}
