% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimilarity.R
\name{cosim_fig}
\alias{cosim_fig}
\title{barpot of the cosinus similarity
barpot of the cosinus similarity}
\usage{
cosim_fig(cosim_matrix, var)
}
\arguments{
\item{cosim_matrix}{the cosinus similarity matrix from cosim_matrix()}

\item{var}{the target variable}
}
\value{
a ggplot barplot
}
\description{
barpot of the cosinus similarity
barpot of the cosinus similarity
}
\author{
Manuel Betin
}
