% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{country_radar_fig}
\alias{country_radar_fig}
\title{depreciated use radar_shocks_fig()
depreciated use radar_shocks_fig()}
\usage{
country_radar_fig(radar_dt)
}
\arguments{
\item{radar_dt}{output of country_radar_fig}
}
\value{
a plotly object
}
\description{
depreciated use radar_shocks_fig()
depreciated use radar_shocks_fig()
}
\author{
Manuel Betin
}
