\name{download_pdf_from_url}
\alias{download_pdf_from_url}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
download a list of pdf files from a list of url paths
}
\description{
download a list of pdf files from a list of url paths
}
\usage{
download_pdf_from_url(urls,export_path,filename=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{urls}{
  list of urls of pdf documents
}
 \item{export_path}{
 path where to download the files
}
\item{filename}{
Name of the files, each element of the file will receive a number
}

\details{

}
\value{
}
\references{
}
\author{
}
\note{

}

\seealso{
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }% use one of  RShowDoc("KEYWORDS")
\keyword{ }% __ONLY ONE__ keyword per line
