% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{find_associated_keywords}
\alias{find_associated_keywords}
\title{depreciated use lexicon_details()
depreciated use lexicon_details()}
\usage{
find_associated_keywords(keyword)
}
\arguments{
\item{keyword}{the name of the category of which you want to 
know the lexicon}
}
\value{
a vector of words
}
\description{
depreciated use lexicon_details()
depreciated use lexicon_details()
}
\author{
Manuel Betin
}
