% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{find_pages}
\alias{find_pages}
\title{depreciated use get_pages()
depreciated use get_pages()}
\usage{
find_pages(file, targetword)
}
\arguments{
\item{file}{a character string correspond to the text
to analysis}

\item{targetword}{a vector of characters corresponding to word to
search and count for in the text}
}
\value{
a list of strings of characters where words have been found.
}
\description{
depreciated use get_pages()
depreciated use get_pages()
}
\author{
Manuel Betin
}
