% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccd.R
\name{get_ccd}
\alias{get_ccd}
\title{Get data from the NCES's Common Core of Data for year 2009 to 2016}
\usage{
get_ccd(states = "all", endyear = 2016, variables = "all")
}
\arguments{
\item{states}{A string of state names (or name) indicating the states that you want data for.}

\item{endyear}{A numeric year indicated the end year of a school year (e.g., 2016 to get 2015-2016 data).}

\item{variables}{A string of variables names for the variables to access. Use "all" for all variables.}
}
\value{
A tibble with school-level data for the given year and state(s)
}
\description{
Get data from the NCES's Common Core of Data for year 2009 to 2016
}
\examples{
# Get data from all schools in New York state for the 2014-2015 school year
library(ccdr)
ny_dat <- get_ccd(states = "New York", endyear = 2015, variables = "all")
}
