% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccd_vars.R
\name{get_ccd_vars}
\alias{get_ccd_vars}
\title{Get variable names from the NCES's Common Core of Data for year 2009 to 2016}
\usage{
get_ccd_vars(endyear = 2016, View = F)
}
\arguments{
\item{endyear}{A numeric year indicated the end year of a school year (e.g., 2016 to get 2015-2016 data).}

\item{View}{Do you want to View the data frame in the R data viewer}
}
\description{
Get variable names from the NCES's Common Core of Data for year 2009 to 2016
}
\examples{
# Get variables from all schools in New York state for the 2014-2015 school year.
ny_vars <- get_ccd_vars(endyear = 2015, View = F)
}
