% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{get_imf_country_reports}
\alias{get_imf_country_reports}
\title{download the url of the country reports from the imf website}
\usage{
get_imf_country_reports(npages = NULL)
}
\arguments{
\item{npages}{the number of pages to download by default it wil download all the pages 
that he will find. to download only the most recent set npages=1}
}
\value{
dataframe with the title of the document, the name of the file and the url of the file
}
\description{
Provide the urls to download the imf country reports from
https://www.imf.org/en/Publications/CR/
}
\author{
Manuel Betin
}
