% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{get_pages}
\alias{get_pages}
\title{Find the page where words a located
Provide files either pdf of html and return the paragraphs matching the
targetted word}
\usage{
get_pages(file, targetword)
}
\arguments{
\item{file}{a character string correspond to the text
to analysis}

\item{targetword}{a vector of characters corresponding to word to
search and count for in the text}
}
\value{
a list of strings of characters where words have been found.
}
\description{
Find the page where words a located
Provide files either pdf of html and return the paragraphs matching the
targetted word
}
\author{
Manuel Betin
}
