% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{idf}
\alias{idf}
\title{Compute the inverse document frequency}
\usage{
idf(tf_data, group = NULL)
}
\arguments{
\item{tf_data}{a dataframe of term frequencies}

\item{group}{the grouping variable}
}
\value{
a tibble with the idf of each categories
}
\description{
The idf is computed as the logarithm of the inverse of the proportion. 
it allows to give reduce weight of words with high
frequency in the corpus
}
\author{
Manuel Betin
}
