% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{idf_barplot}
\alias{idf_barplot}
\title{barplot for the idf
barplot displaying the idf of the categories}
\usage{
idf_barplot(idf, vars_type = c("economic_shock", "non_economic_shock",
  "debt_outcomes", "debt_structure", "characteristics_program",
  "adustment_program"), vars_nature = NULL, idf_trans = F)
}
\arguments{
\item{idf}{a tibble of idf as produced by idf()}

\item{vars_type}{a vector with the type of shocks to include (see lexicon_typology() for classification)}

\item{vars_nature}{a vector of the nature of shocks to include (see lexicon_typology() for classification)}

\item{idf_trans}{T use the term frequency transformation, F use the proportion}
}
\value{
a ggplot barplot object
}
\description{
barplot for the idf
barplot displaying the idf of the categories
}
\author{
Manuel Betin
}
