% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{key_word_crisis}
\alias{key_word_crisis}
\title{depreciated: use lexicon(),
depreciated: use lexicon(),}
\usage{
key_word_crisis()
}
\value{
a list of categories with the corresponding words
 associated
}
\description{
depreciated: use lexicon(),
depreciated: use lexicon(),
}
