\name{key_words_categories}
\alias{key_words_categories}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 generate a lexicon of categories of crisis as alist of key words by category of capital letter do not matter
}
\description{
 generate a lexicon of categories of crisis as alist of key words by category
}
\usage{
key_words_categories()
}
\author{
Manuel Betin
}