% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.r
\name{lexicon_details}
\alias{lexicon_details}
\title{Provide the keywords associate to the name of the category}
\usage{
lexicon_details(keyword)
}
\arguments{
\item{keyword}{the name of the category of which you want to
know the lexicon}
}
\value{
a vector of words
}
\description{
usefull know what are the words behind each category
}
\author{
Manuel Betin
}
