% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.r
\name{lexicon_typology}
\alias{lexicon_typology}
\title{provide the typology of each category}
\usage{
lexicon_typology()
}
\value{
a dataframe classifying each category of the lexicon
}
\description{
provide the typology of each category
}
\author{
Manuel Betin
}
