% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{log_norm_trans}
\alias{log_norm_trans}
\title{log normal transformation of the table
log normal transformation of the table}
\usage{
log_norm_trans(tf_data)
}
\arguments{
\item{tf_data}{a dataframe of term frequencies}
}
\value{
a tibble of tf with log norm transformation
}
\description{
log normal transformation of the table
log normal transformation of the table
}
\author{
Manuel Betin
}
