% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{pdf_page_count}
\alias{pdf_page_count}
\title{count the number of pages in the pdf
count the number of pages in the pdf}
\usage{
pdf_page_count(files)
}
\arguments{
\item{files}{a list of character strings}
}
\value{
the number of pages in the document
}
\description{
count the number of pages in the pdf
count the number of pages in the pdf
}
\author{
Manuel Betin
}
