\name{plot_cos_sim}
\alias{plot_cos_sim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute the cosinus similarity matrix by comparing pairwise columns
}
\description{
Compute the cosinus similarity matrix by comparing pairwise columns
}
\usage{
plot_cos_sim(cos_sim, var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cos_sim}{
cosinus similarity matrix output of matrix_cosinus_similarity()
}
  \item{var}{
the name of the category to plot
}
}
\value{
plot results of matric cosinus for one selected column 
}
\author{
Manuel Betin
}