% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.corpusTM}
\alias{print.corpusTM}
\title{Printing CorpusTM
S3 method for class "corpus.TM"}
\usage{
\method{print}{corpusTM}(x, ...)
}
\arguments{
\item{x}{object of class "corpus.TM"}

\item{...}{optional arguments to ```print```}
}
\description{
Printing CorpusTM
S3 method for class "corpus.TM"
}
\author{
Manuel Betin, Umberto Collodel
}
