% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar_dt}
\alias{radar_dt}
\title{Transform the tf-idf database in proper format for drawing the radar chart of categories
Transform the tf-idf database in proper format for drawing the radar chart of categories}
\usage{
radar_dt(tf_data, isoc, top_n = 50, weight_method = "brut_frequency",
  group = "ISO3_Code")
}
\arguments{
\item{tf_data}{a dataframe with documents in rows and tf-idfs of each categories
in columns}

\item{isoc}{the iso3 code of the country to display}

\item{top_n}{the first n larger index to display}

\item{weight_method}{the type of method for the tf computation one of "brut_frequency",
"binary_frequency","log_norm_frequency"}

\item{group}{the grouping variable of interest}
}
\value{
a tible with the value of the tf summarized according to the grouping selected
}
\description{
Transform the tf-idf database in proper format for drawing the radar chart of categories
Transform the tf-idf database in proper format for drawing the radar chart of categories
}
\author{
Manuel Betin
}
