% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar_reforms_fig}
\alias{radar_reforms_fig}
\title{Radar chart of the tf}
\usage{
radar_reforms_fig(radar_dt)
}
\arguments{
\item{radar_dt}{output of country_radar_fig}
}
\value{
a plotly object
}
\description{
Use the radar_dt() output and draw a radar chart to display the 
term frequencies of the reforms categories.
}
\author{
Manuel Betin
}
