% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar_shocks_fig}
\alias{radar_shocks_fig}
\title{Radar chart of the tf}
\usage{
radar_shocks_fig(radar_dt)
}
\arguments{
\item{radar_dt}{output of country_radar_fig}
}
\value{
a plotly object
}
\description{
Use the radar_dt() output and draw a radar chart to display the 
term frequencies of the selected categories. Indexes are order counterclockwise
from more exogeneous (15:15) to more exogeneous
}
\author{
Manuel Betin
}
