\name{remove_stop}
\alias{remove_stop}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Remove stop words from a string of characters
}
\description{
Remove stop words from a string of characters
}
\usage{
remove_stop(word)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{word}{
string of character
}
}
\value{
string of characters without the stop words
}
\author{
Manuel Betin
}
