% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{run_tf}
\alias{run_tf}
\title{Compute the tf matrix for a corpus given in a list format,}
\usage{
run_tf(corpus_path = "IMF_letofIntent_1960_2014_clean.RData",
  type_lexicon = "words", keyword_list = c("Commodity_crisis",
  "Balance_payment_crisis", "Inflation_crisis"), export_path = NULL,
  parrallel = T)
}
\arguments{
\item{corpus_path}{the path to the RData file containing the corpus to analyze}

\item{type_lexicon}{"words" or "categories" for the type of lexicon to use}

\item{keyword_list}{the categories on which computing the term frequency}

\item{export_path}{the path were to export the tf}

\item{parrallel}{T/F to use mclapply from the parrallel package}
}
\value{
a dataframe of term frequencies
}
\description{
the type of lexicon to use and the sublist of keywords associated The
output is a matrix of tf with a row per document and a column for each
element of the keyword list
}
\author{
Manuel Betin
}
