% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{run_tf_by_chunk}
\alias{run_tf_by_chunk}
\title{run the term frequency matrix on the list of urls provided as parameter.
The function download the pdf, create the corpus and generate the term
frequency matrix, to avoid storage limitation the function deleate
original pdf files}
\usage{
run_tf_by_chunk(urls = url_links, keyword_list = c("Fiscal outcomes",
  "Currency_crisis"), extract_number = 1, delete_pdfs = T,
  rm_short_docs = F, min_words = 100, parrallel = T,
  loc_temp = NULL)
}
\arguments{
\item{urls}{a dataframe containing two columns: name_file and pdf
that respectively provid the name that will be given to the file
downloaded and pdf that contains the url of the pdf}

\item{keyword_list}{one of the element of the list provided by
lexicon()}

\item{extract_number}{a number that will be ued as a suffix for the name 
of the corpus and the name of the tf matrix}

\item{delete_pdfs}{T/F, set to T it will delete the folder 
containing the original pdf, usefull option when the number of pdf is very
large and the size of the folder start to be very large}

\item{rm_short_docs}{T/F T if you want to remove the documents under a 
certain number of words}

\item{min_words}{the minimum word in the document necessary to perform
the text mining}

\item{parrallel}{if T then will use mclapply from the package parrallel}

\item{loc_temp}{path to the folder containing the files, corpus and tf}
}
\value{
create forlders where pdfs, corpus and tf-idf matrix are stored
}
\description{
run the term frequency matrix on the list of urls provided as parameter.
The function download the pdf, create the corpus and generate the term
frequency matrix, to avoid storage limitation the function deleate
original pdf files
}
\author{
Manuel Betin
}
