% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{run_tf_update}
\alias{run_tf_update}
\title{Updates the tf-indexes without repeating full extraction}
\usage{
run_tf_update(
  path_tf_to_update = "tf_crisis_words.RData",
  corpus_path = "IMF_letofIntent_1960_2014_clean.RData",
  type_lexicon = "words",
  keyword_list = NULL,
  export_path = "tf_crisis_words.RData",
  parrallel = T,
  store_old = F,
  store_old_path = NULL
)
}
\arguments{
\item{path_tf_to_update}{Path to old tf dataframe.}

\item{corpus_path}{Path to corpus from which perform the extraction.}

\item{type_lexicon}{Character: "words" or "category"}

\item{keyword_list}{Character vector: names of character vectors to use for extraction.}

\item{export_path}{Path to export the file.}

\item{parrallel}{Logical. If TRUE, parallel computation for each category.}

\item{store_old}{Logical. If TRUE, store old extractions in a directory.}

\item{store_old_path}{Path to move old extractions in.}
}
\value{
A dataframe with file name, old indexes not specified in keyword_list and updated indexes.
}
\description{
Combine probability of shocks, intensity and complexity of relations to construct
.a mesure of severity of crisis
}
\examples{


}
\author{
Manuel Betin, Umberto Collodel
}
