% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{scrap.ccdr.files}
\alias{scrap.ccdr.files}
\title{download pdf documents
download from a a dataframe containing the url of the files
the pdf of interest}
\usage{
scrap.ccdr.files(urls, export_path, overwrite = T)
}
\arguments{
\item{urls}{a dataframe with a row by document and at least the following columns
title, pdf, and name_file. title contains the title of the document, pdf the url
where to download the files and name_file the name that will be given to the pdf}

\item{export_path}{path where to store the downloaded pdfs}

\item{overwrite}{T the function re download the file and overwrite the existing}
}
\value{
creates a folder and store the pdf
}
\description{
download pdf documents
download from a a dataframe containing the url of the files
the pdf of interest
}
\author{
Manuel Betin
}
