\name{search_contagion}
\alias{search_contagion}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
provide the keywords associated to each category of crisis
}
\description{
provide the keywords associated to each category of crisis
}
\usage{
search_contagion(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
a character string where to check for mentions of other countries 
}
}
\value{
the function returns a dataframe with the country names and mnemonics
}
\author{
Manuel Betin
}