\name{TextMiningCrisis-package}
\alias{TextMiningCrisis-package}
\alias{TextMiningCrisis}
\docType{package}
\title{
\packageTitle{TextMiningCrisis}
}
\description{
\packageDescription{TextMiningCrisis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TextMiningCrisis}
\packageIndices{TextMiningCrisis}
 
Functions documented:

 1. key_words_crisis 
 2. key_words_category
 3. remove_stop
 4. ngram_clean
 5. associated_words
 6. Unsupervized_corpus
 7. find_pertinance
 8. clean_text     
 9. Find_targeted_pages
 10. Eval_targeted_pages
 11. table_brute_frequency
 12. table_brute_frequency_vector
 13. binary_frequency_trans
 14. log_norm_trans
 15. invers_doc_frequency
 16. table_word_weight
 17. run_tf
 18. aggregate_corpus
 19. cosinus_similarity
 20. plot_cos_sim
 21. evol_weight_by_group
 22. country_radar_dt
 23. country_radar_fig
 24. find_associated_keyword
 25. keywords_country_contagion
 26. search_contagion
}
\author{
\packageAuthor{TextMiningCrisis}

Maintainer: \packageMaintainer{TextMiningCrisis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ Text Mining, Economic crisis, Natural lenguage processing }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
