% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{tf_barplot}
\alias{tf_barplot}
\title{barplot of term frequencies
take as input the term frequency matrix and reshape it to plot the values
for all numeric columns and group them in two dimensions: the type of
variable and the nature of the shock (exogeneous or endogenous)}
\usage{
tf_barplot(tf, vars_type = c("economic_shock", "non_economic_shock",
  "debt_outcomes", "debt_structure", "characteristics_program",
  "adustment_program"), vars_nature = NULL)
}
\arguments{
\item{tf}{a term frequency matrix from tf()}

\item{vars_type}{the type of categories to include (see lexicon_typology()
for more details)}

\item{vars_nature}{the nature of categories to include (see lexicon_typology()
for more details)
@author Manuel Betin
@return a ggplot barplot
@export}
}
\description{
barplot of term frequencies
take as input the term frequency matrix and reshape it to plot the values
for all numeric columns and group them in two dimensions: the type of
variable and the nature of the shock (exogeneous or endogenous)
}
