% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{tf_by_group}
\alias{tf_by_group}
\title{Summarize the term frequency matrix by group
Summarize the tf matrix according to the grouping variable selected}
\usage{
tf_by_group(tf_data, weight_method = "brut_frequency", mygroup = NULL)
}
\arguments{
\item{tf_data}{a tibble of term frequencies from tf(), run_tf(),
run_tf_update() or run_tf_by_chunk()}

\item{weight_method}{the method for countring "brut_frequency", "binary_frequency"}

\item{mygroup}{the grouping variable to be used for the summary}
}
\value{
a dataframe of term frequencies with documents in rows and categories
in columns
}
\description{
Summarize the term frequency matrix by group
Summarize the tf matrix according to the grouping variable selected
}
\author{
Manuel Betin
}
