\name{typology_categories}
\alias{typology_categories}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Provide a dataframe with the typology and nature of the shock associated to each category
}
\description{
Provide a dataframe with the typology and nature of the shock associated to each category
}
\usage{
typology_categories()
}
%- maybe also 'usage' for other objects documented here.
\author{
Manuel Betin
}