\name{unsupervized_corpus}
\alias{unsupervized_corpus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
from the table of ngrams generated by ngram_clean() create the corpus of documents
}
\description{
from the table of ngrams generated by ngram_clean() create the corpus of documents
}
\usage{
unsupervized_corpus(corpus, length_ngram = 2, min_occurence = 10)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{corpus}{
 the list of files to analyse
}
  \item{length_ngram}{
the length of the ngram considered
}
  \item{min_occurence}{
 minimum occurrence to keep
}
}
\value{
a tibble with ngrams in rows, number of occurence and dummies for the typologies 
}
\author{
Manuel Betin
}