\name{CCDr-package}
\alias{CCDr-package}
\alias{CCDr}
\docType{package}
\title{
\packageTitle{CCDr}
}
\description{
\packageDescription{CCDr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CCDr}
\packageIndices{CCDr}
 
}
\author{
\packageAuthor{CCDr}

Maintainer: \packageMaintainer{CCDr}
}
\references{
- Manuel Bétin, Umberto Collodel. The Complex Crises Database: 70 Years of Macroeconomic Crises. 2021. https://halshs.archives-ouvertes.fr/halshs-03268889/

- Link to the database: 10.7910/DVN/Y8CN0A


}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{ Text Mining, Economic crisis, Natural lenguage processing}