% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.r
\name{ccdr.fasttext.dataforprediction.clean}
\alias{ccdr.fasttext.dataforprediction.clean}
\title{preprocess the sentences and prepare for prediction}
\usage{
ccdr.fasttext.dataforprediction.clean(myfile, rm_stopwords = T)
}
\arguments{
\item{myfile}{}

\item{rm_stopwords}{T/F whether remove stopwords and numbers in the predictions}
}
\description{
clean the sentences by removing numbers,
short sentences, irrelevant sentences and stopwords
}
\author{
manuel betin
}
