% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{ccdr.tfs.fig.radiography}
\alias{ccdr.tfs.fig.radiography}
\title{produce radiography of crisis}
\usage{
ccdr.tfs.fig.radiography(CCDB_y, iso3c = "all", vars, years_crisis, window)
}
\arguments{
\item{CCDB_y}{a vintage of the CCDB database}

\item{iso3c}{a iso3 code of country}

\item{vars}{a vector of crisis names available in the CCDB_y dataset}

\item{years_crisis}{a vector of crisis dates}

\item{window}{a vector of number of years before and after the crisis to display}
}
\description{
figure that create the radiography of crisis for the selected 
variables and period window
}
\author{
Manuel Betin
}
