% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaning.R
\name{clean_text}
\alias{clean_text}
\title{clean character strings
clean character string by removing special characters 
and irrelevant strings}
\usage{
clean_text(file)
}
\arguments{
\item{file}{a string of characters to clean}
}
\value{
a clean character string
}
\description{
clean character strings
clean character string by removing special characters 
and irrelevant strings
}
\author{
Manuel Betin
}
