\name{cosim}
\alias{cosim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
formula for the cosinus similiarity that provides the distance between two vectors here two crisis
}
\description{
formula for the cosinus similiarity that provides the distance between two vectors here two crisis
}
\usage{
cosim(table_N_occurence, vec1, vec2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{table_N_occurence}{
output of table_brute_frequency_vector()
}
  \item{vec1}{
vector of tf.idf
}
  \item{vec2}{
vector of tf.idf
}
}
\value{
the function returns a numeric value corresponding to the cosinus similarity
}
\author{
Manuel Betin
}