\name{cosim_matrix}
\alias{cosim_matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute the cosinus similarity matrix by comparing pairwise columns
}
\description{
Compute the cosinus similarity matrix by comparing pairwise columns
}
\usage{
cosim_matrix(table_weights)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{table_weights}{
tf.idf dataframe
}
}
\author{
Manuel Betin
}


