% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{eval_pages}
\alias{eval_pages}
\title{Look for the presence of the targetword into a character string}
\usage{
eval_pages(files, targetword, brute_freq = F, parrallel = T)
}
\arguments{
\item{files}{a character string correspond to the text to
analysis}

\item{targetword}{a vector of characters corresponding to word to
search and count for in the text}

\item{brute_freq}{T if you only want the frequency count of the word,
if F the output is the it provides the term frequency for the set of 
targetwords provided}

\item{parrallel}{if T the function use the mclapply function to use
parrallele computing from the package "parrallele"
@author Manuel Betin
@return  return the number of occurence of the targetword for each file 
@export}
}
\description{
Provide list of files and return summary of counts of occurence of the
target world
}
