\name{Key_words_crisis}
\alias{Key_words_crisis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
generate a lexicon of categories of crisis as a list of key words by category
}
\description{
 generate a lexicon of categories of crisis as a  list of key words by category

}
\usage{
Key_words_crisis()
}
%- maybe also 'usage' for other objects documented here.
\author{
Manuel Beitn
}