\name{log_norm_trans}
\alias{log_norm_trans}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 log normal transformation of the table
}
\description{
 log normal transformation of the table
}
\usage{
log_norm_trans(table_N_occurence)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{table_N_occurence}{
output of table_brute_frequency_vector()
}
}
\value{
the function returns the log normal transformation of the dataframe.
}

\author{
Manuel Betin
}