\name{pdf_page_count}
\alias{pdf_page_count}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
count the number of pages in the pdf
}
\description{
count the number of pages in the pdf
}
\usage{
pdf_page_count(files)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{files}{
files: a list of character strings
}
}
\author{
Manuel Betin
}