% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{plot_cos_sim}
\alias{plot_cos_sim}
\title{depreciated use cosim_fig()
depreciated use cosim_fig()}
\usage{
plot_cos_sim(cosim_matrix, var)
}
\arguments{
\item{cosim_matrix}{the cosinus similarity matrix from cosim_matrix()}

\item{var}{the target variable}
}
\value{
a ggplot barplot
}
\description{
depreciated use cosim_fig()
depreciated use cosim_fig()
}
\author{
Manuel Betin
}
