\name{radar_reforms_fig}
\alias{radar_reforms_fig}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Look at average importance of each reform and display the more important elements
}
\description{
Look at average importance of each reform by country and display the more important elements
}
\usage{
country_radat_fig(country_radar_dt)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{country_radar_dt}{
output of country_radar_dt()
}
}
\value{
the function returns a radar chart showing the data of country_radar_dt
}
\author{
Manuel Beitn
}