% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radar_structure_fig}
\alias{radar_structure_fig}
\title{Radar chart of the tf 
Use the radar_dt() output and draw a radar chart to display the 
term frequencies of the debt structure category.}
\usage{
radar_structure_fig(radar_dt)
}
\arguments{
\item{radar_dt}{output of country_radar_fig}
}
\value{
a plotly object
}
\description{
Radar chart of the tf 
Use the radar_dt() output and draw a radar chart to display the 
term frequencies of the debt structure category.
}
\author{
Manuel Betin
}
