% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{run.ccdr.tfs}
\alias{run.ccdr.tfs}
\title{Compute the tf matrix for a corpus given in a list format,}
\usage{
run.ccdr.tfs(corpus_file, lexicon, export_path = NULL, parrallel = T)
}
\arguments{
\item{corpus_file}{the path to the RData file containing the corpus to analyze}

\item{lexicon}{a lexicon in the form of a list containing a vector of keywords, 
the list is named with the name of the category of interest}

\item{export_path}{the path were to export the tf}

\item{parrallel}{T/F to use mclapply from the parrallel package}
}
\value{
a dataframe of term frequencies
}
\description{
the type of lexicon to use and the sublist of keywords associated The
output is a matrix of tf with a row per document and a column for each
element of the keyword list
}
\author{
Manuel Betin
}
