\name{TextMiningCrisis-package}
\alias{TextMiningCrisis-package}
\alias{TextMiningCrisis}
\docType{package}
\title{
\packageTitle{TextMiningCrisis}
}
\description{
\packageDescription{TextMiningCrisis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TextMiningCrisis}
\packageIndices{TextMiningCrisis}
 
}
\author{
\packageAuthor{TextMiningCrisis}

Maintainer: \packageMaintainer{TextMiningCrisis}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{ Text Mining, Economic crisis, Natural lenguage processing }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
