% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{tf}
\alias{tf}
\title{Compute the term freqeuency matrix
for each document in the corpus the function creates a table counting the
occurence of the words in the vector of keywords and provide a table with
the number of occurence, name of the file that is the name of each
document in the corpus}
\usage{
tf(corpus, keywords, brute_freq = F, parrallel = T)
}
\arguments{
\item{corpus}{a list of texts from pdf_from_url() with different names
for each element}

\item{keywords}{a vector of strings containing the targeted
words to look for}

\item{brute_freq}{T/F if set to F it will compute the term frequency
otherwise it will only count the occurence}

\item{parrallel}{T/F if T use mclapply from parrallel package}
}
\value{
a dataframe of term frequencies with documents in rows and categories
in columns
}
\description{
Compute the term freqeuency matrix
for each document in the corpus the function creates a table counting the
occurence of the words in the vector of keywords and provide a table with
the number of occurence, name of the file that is the name of each
document in the corpus
}
\author{
Manuel Betin
}
