% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{tf_idf}
\alias{tf_idf}
\title{Compute the tf-idf matrix}
\usage{
tf_idf(tf_data, weight_method = "brut_frequency")
}
\arguments{
\item{tf_data}{the tf matrix from tf()}

\item{weight_method}{the method for the counting: "binary_frequency" or "brut_frequency"}
}
\value{
a dataframe of tf-idf with documents in rows and categories
in columns
}
\description{
weight the term frequency (tf) of each category by the its idf to
improve the discriminatory power of the categories that are rare in
the corpus
}
\author{
Manuel Betin
}
