% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_frequencies.R
\name{tf_vector}
\alias{tf_vector}
\title{vectorize the function tf() to be able to pass a list of names of keywords
keyword_list is a list containing the names of different groups of
keywords that have a vector of words to look into.}
\usage{
tf_vector(corpus, keyword_list, brute_freq = F, parrallel = T,
  centre_countries = "USA")
}
\arguments{
\item{corpus}{a list of texts from pdf_text() with different names
for each element}

\item{keyword_list}{the names of the items in lexicon to
include in the computation}

\item{brute_freq}{T/F if T it will just count the occurence, otherwise
it will compute the term frequency}

\item{parrallel}{T/F if T it will use mclapply from the parrallel package}

\item{centre_countries}{Character string. Default is "USA". Netting of confusing
keywords will not be performed countries in the vector.
 @author Manuel Betin
 @return a tibble with the term frequencies for the selected categories
 @export}
}
\description{
vectorize the function tf() to be able to pass a list of names of keywords
keyword_list is a list containing the names of different groups of
keywords that have a vector of words to look into.
}
