% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{typology_categories}
\alias{typology_categories}
\title{depreciated use lexicon_typology()
depreciated use lexicon_typology()}
\usage{
typology_categories()
}
\value{
a dataframe classifying each category of the lexicon
}
\description{
depreciated use lexicon_typology()
depreciated use lexicon_typology()
}
\author{
Manuel Betin
}
